\name{campaignmonitoR-package}
\alias{campaignmonitoR-package}
\alias{campaignmonitoR}
\docType{package}
\title{
\packageTitle{campaignmonitoR}
}
\description{
\packageDescription{campaignmonitoR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{campaignmonitoR}
\packageIndices{campaignmonitoR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{campaignmonitoR}

Maintainer: \packageMaintainer{campaignmonitoR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
