\name{campaignmonitoR-package}
\alias{campaignmonitoR-package}
\alias{campaignmonitoR}
\docType{package}
\title{
\packageTitle{campaignmonitoR}
}
\description{
\packageDescription{campaignmonitoR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{campaignmonitoR}
\packageIndices{campaignmonitoR}
This package is designed to interface with the campaign monitor API in an intuitive manner to assist with bespoke analysis of email marketing activity.  
}
\author{
\packageAuthor{campaignmonitoR}

Maintainer: \packageMaintainer{campaignmonitoR}
}
\references{
Christoph Waldhauser. 2015. Accessing APIs from R (and a little R programming). [ONLINE] Available at: https://www.r-bloggers.com/accessing-apis-from-r-and-a-little-r-programming/. [Accessed 29 June 2018].
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ TODO ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ TODO ~~
}
